+++ CommNetAntennasInfo/Properties/AssemblyVersion.cs
@@ -4,4 +4,4 @@
   
  using System.Reflection;
 
- [assembly: AssemblyVersion("2.1.0.0")]
+ [assembly: AssemblyVersion("2.1.1.0")]