+++ GameData/CommNetAntennasInfo/ChangeLog.md
@@ -1,5 +1,11 @@
 ChangeLog:
 
+## Version 2.1.1
+ * PAW: fix default rating
+ * PAW: compact for internal antenna, 
+ * update MM 4.0.3
+
+
 ## Version 2.1.0
  * Antenna Type and CombinabilityExponent in the PAW (Part-Action-Window)
    It show up only with enabled Advanced Tweakable setting